+++ react-todo/src/components/templates/TodoCreateTemplate/useTodoCreateTemplate.ts
@@ -1,0 +1,60 @@
+import { useCallback, useState } from 'react';
+import { useNavigate } from 'react-router-dom';
+import { NAVIGATION_PATH } from '../../../constants/navigation';
+import { EventType } from '../../../interfaces/Event';
+
+type Param = {
+  addTodo: (title: string, content: string) => void;
+};
+
+type StatesType = {
+  inputTitle: string;
+  inputContent: string;
+};
+
+type ActionsType = {
+  handleChangeTitle: EventType['onChangeInput'];
+  handleChangeContent: EventType['onChangeTextArea'];
+  handleCreateTodo: EventType['onSubmit'];
+};
+
+export const useTodoCreateTemplate = ({ addTodo }: Param) => {
+  const navigate = useNavigate();
+
+  const [inputTitle, setInputTitle] = useState('');
+  const [inputContent, setInputContent] = useState('');
+
+  const handleChangeTitle: EventType['onChangeInput'] = useCallback(
+    (e) => setInputTitle(e.target.value),
+    [],
+  );
+
+  const handleChangeContent: EventType['onChangeTextArea'] = useCallback(
+    (e) => setInputContent(e.target.value),
+    [],
+  );
+
+  const handleCreateTodo: EventType['onSubmit'] = useCallback(
+    (e) => {
+      e.preventDefault();
+      if (inputTitle !== '' && inputContent !== '') {
+        addTodo(inputTitle, inputContent);
+        navigate(NAVIGATION_PATH.TOP);
+      }
+    },
+    [addTodo, inputTitle, inputContent, navigate],
+  );
+
+  const states: StatesType = {
+    inputTitle,
+    inputContent,
+  };
+
+  const actions: ActionsType = {
+    handleChangeTitle,
+    handleChangeContent,
+    handleCreateTodo,
+  };
+
+  return [states, actions] as const;
+};
